Code-format and finish renaming DataFormats folder
